feat(ecs): Phase 1+2 prerequisites, BitMask512, EntityMetadataCold (BATCH-01)
TASK-E001: Widen component ID type byte->int
- ComponentIdAttribute.Id: byte -> int; constructor parameter widened
- GlobalComponentIds: all const byte -> const int (values unchanged)
- ID block 256-511 reserved for future components

TASK-E002: Engine capacity constants + QueryBuilder guard
- FdpConfig.MAX_COMPONENT_TYPES: 256 -> 512
- FdpConfig.FORMAT_VERSION: 4 -> 5
- QueryBuilder/EntityQuery internal masks upgraded from BitMask256 to
  BitMask512 to satisfy WithComponentId(256-511) test requirement

TASK-E003: New BitMask512 (64 bytes / one L1 cache line)
- 8 ulong fields at LayoutKind.Explicit offsets 0-56
- AVX2 two-stage path: lower 256 bits first, upper 256 bits second
- Scalar fallback: interleaved include/exclude checks per quad
- FDP_PARANOID_MODE bounds guards for SetBit/ClearBit/IsSet
- Phase-2 compatibility overloads HasAll/HasAny(BitMask256, BitMask512)

TASK-E004: New EntityMetadataCold (128 bytes / 2 cache lines)
- AuthorityMask (BitMask512 at offset 0), Generation, Flags,
  LastChangeTick, DisType, LifecycleState
- IsActive/SetActive operate on Flags bit 0 without touching other bits

Tests: 764 passed, 2 skipped, 0 failed

+++ FDP/Engine/Fdp.Core/ComponentIdAttribute.cs
@@ -17,7 +17,7 @@ namespace Fdp.Core
     ///
     /// <para>
     /// IDs are allocated in named blocks defined by <see cref="GlobalComponentIds"/>.
-    /// Component IDs are limited to the range [0, 255] by the <c>BitMask256</c> capacity.
+    /// Component IDs are limited to the range [0, 511] by the <c>BitMask512</c> capacity.
     /// </para>
     /// </summary>
     /// <example>
@@ -30,20 +30,20 @@ namespace Fdp.Core
     public sealed class ComponentIdAttribute : Attribute
     {
         /// <summary>
-        /// Stable, globally unique component type ID in the range [0, 255].
+        /// Stable, globally unique component type ID in the range [0, 511].
         /// Must match the corresponding constant in <see cref="GlobalComponentIds"/>.
         /// </summary>
-        public byte Id { get; }
+        public int Id { get; }
 
         /// <summary>
         /// Initialises the attribute with a stable component ID.
         /// </summary>
         /// <param name="id">
-        /// The globally unique component ID (0–255). Must not collide with any other
+        /// The globally unique component ID (0–511). Must not collide with any other
         /// component's ID anywhere in the codebase; a collision throws
         /// <see cref="InvalidOperationException"/> at runtime during registration.
         /// </param>
-        public ComponentIdAttribute(byte id)
+        public ComponentIdAttribute(int id)
         {
             Id = id;
         }
